add user routes, reducer and actions

+++ api/src/routes/favorites/controller.ts
@@ -1,17 +1,16 @@
 import {Response, Request} from 'express';
 import FavoritesModels from '../../models/Favorites';
 
-
 /////// Put ////////
 export const addFavorites = async (req:Request ,res:Response) =>{
-    const idUser:string = req.body.idUser;
-    const city:string = req.body.city;
+    const city:string = req.body.city; 
+    const idUser = req.user;
     try{
         if(idUser) {
             const favorite = await FavoritesModels.findOne({user:idUser})
-            if(!favorite?.favorites.find(el => el === city)){
+            if(!favorite?.favorites.find(el => el === city && favorite?.favorites.length <= 5)){
             favorite?.favorites.push(city)
-             await FavoritesModels.updateOne({user:idUser},{favorites:favorite?.favorites})
+            await FavoritesModels.updateOne({user:idUser},{favorites:favorite?.favorites})
             res.status(200).send('successfully add')
             }else if(favorite?.favorites.find(el => el === city)){
                 res.status(404).json({data:'the city already exists in your favorites'})
@@ -24,8 +23,8 @@ export const addFavorites = async (req:Request ,res:Response) =>{
 };
 
 export const removeFavorites = async (req:Request ,res:Response) =>{
-    const idUser:string = req.body.idUser;
     const city:string = req.body.city;
+    const idUser = req.user;
     try{
         if(idUser) {
             const favorite = await FavoritesModels.findOne({user:idUser})
@@ -39,15 +38,16 @@ export const removeFavorites = async (req:Request ,res:Response) =>{
     }
 } ;
 
-
-
-
 ///////// delete ///////////
 export const deleteFavorites = async (req:Request ,res:Response) =>{
-const idUser:string = req.body.idUser;
+    const idUser = req.user;
     try{
+        if(idUser){
         await FavoritesModels.deleteOne({user:idUser}) 
         res.status(200).send('successfully deleted')
+        return;
+        };
+        res.status(404).json({data:'error as ocurrent'})
 }catch(err){
     res.status(404).json({data:err+''})
 };
@@ -55,16 +55,13 @@ const idUser:string = req.body.idUser;
 
 
 
-
-
 /////////// get //////////
 
 export const getFavorites = async (req:Request ,res:Response) =>{
-
-const idUser: string= req.body.idUser;
+    const idUser = req.user;
 try{
     if(idUser){
-    const favorites = await FavoritesModels.findOne({user:idUser},{favorites:1, _id:0});
+    const favorites = await FavoritesModels.findOne({idUser:idUser},{favorites:1, _id:0});
     res.status(200).json(favorites)
     }
     res.status(404).json({data:"No favorites found"})
